+
This paper exploits the concept of nonlinear sliding surfaces to be used as a basis in the development of aerial path planning projects involving aerobatic three-dimensional path curves in the presence of disturbances. This approach can be used for any kind of unmanned aerial vehicle aimed at performing aerobatic maneuvers. Each maneuver is associated with a nonlinear surface on which an aerial vehicle could be driven to slide. The surface design exploits the properties of Viviani’s curve and the Hopf bifurcation. A vector form of the super twisting algorithm steers the vehicle to the prescribed surfaces. A suitable switching control law is proposed to shift between surfaces at different time instants. A practical stability analysis that involves the descriptor approach allows for determining the controller gains. Numerical simulations are developed to illustrate the accomplishment of the suggested aeroba